+++ composition-js/src/__tests__/compose.composeDirective.test.ts
@@ -960,4 +960,38 @@ describe('composing custom core directives', () => {
     const appliedDirectives = schema.elementByCoordinate('Query.shared')?.appliedDirectives;
     expect(appliedDirectives?.map(d => [d.name, d.arguments()])).toMatchObject([['auth', { scope: 'VIEWER'}], ['auth', { scope: 'ADMIN'}]]);
   });
+
+  it('custom directive on nullable array', () => {
+    const subgraphA = {
+      typeDefs: gql`
+        extend schema @composeDirective(name: "@auth")
+          @link(url: "https://specs.apollo.dev/federation/v2.1", import: ["@key", "@composeDirective", "@shareable"])
+          @link(url: "https://custom.dev/auth/v1.0", import: ["@auth"])
+          directive @auth(scope: [String!]) repeatable on FIELD_DEFINITION
+
+        type Query {
+          shared: String @shareable @auth(scope: "VIEWER")
+        }
+      `,
+      name: 'subgraphA',
+    };
+
+    const subgraphB = {
+      typeDefs: gql`
+        extend schema @composeDirective(name: "@auth")
+          @link(url: "https://specs.apollo.dev/federation/v2.1", import: ["@key", "@composeDirective", "@shareable"])
+          @link(url: "https://custom.dev/auth/v1.0", import: ["@auth"])
+          directive @auth(scope: [String!]) repeatable on FIELD_DEFINITION
+
+          type Query {
+          shared: String @shareable @auth
+        }`,
+      name: 'subgraphB',
+    };
+
+    const result = composeServices([subgraphA, subgraphB]);
+    const schema = expectNoErrors(result);
+    const appliedDirectives = schema.elementByCoordinate('Query.shared')?.appliedDirectives;
+    expect(appliedDirectives?.map(d => [d.name, d.arguments()])).toMatchObject([['auth', { scope: 'VIEWER'}], ['auth', {}]]);
+  });
 });